update swets list

+++ src/GiftBox/Marmalade.java
@@ -3,7 +3,7 @@ package GiftBox;
 public class Marmalade extends Sweet {
 
 
-    public Marmalade(String name, Integer count, double price, int weight) {
+    public Marmalade(String name, Integer count, double price, double weight) {
         super(name, count, price, weight);
     }
